+++ .idea/libraries/sqlite_jdbc_3_30_1.xml
@@ -1,7 +1,7 @@
 <component name="libraryTable">
   <library name="sqlite-jdbc-3.30.1">
     <CLASSES>
-      <root url="jar://C:/Program Files/JetBrains/sqlite-jdbc-3.30.1.jar!/" />
+      <root url="jar://$PROJECT_DIR$/../sqlite-jdbc-3.30.1.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES />
