+++ untitled/src/Model/account.java
@@ -1,5 +1,8 @@
 
 package Model;
+
+import java.sql.*;
+
 public class account
 {
 
@@ -16,6 +19,11 @@ public class account
         this.firstname = firstname;
         this.lastname = lastname;//
 
+    }
+
+    public account(String username, String password) {
+        this.username = username;
+        this.password = password;
     }
 
     public String getUsername()
@@ -57,5 +65,36 @@ public class account
 
     public void setLastname(String lastname) {
         this.lastname = lastname;
+    }
+
+    public Boolean ValidateLoginRequest (String Username, String pass) throws SQLException
+    {
+        String windownsUserName = System.getProperty("user.name");
+        try
+        {
+            System.out.println("Connecting to a selected database...");
+            Connection connection = DriverManager.getConnection("jdbc:sqlite:C:\\Users\\"+windownsUserName+"\\IdeaProjects\\Software-engineer-final-project\\untitled\\src\\DButills\\account.db");
+            System.out.println("Connected database successfully...");
+            String sqlQuary = "SELECT username FROM accounts WHERE username = '"+Username+"' AND pass = '"+pass+"'";
+            PreparedStatement ps = connection.prepareStatement(sqlQuary);
+            ResultSet rs = ps.executeQuery();
+            if (rs.next()) {
+                System.out.println("you have successfully logged in to " + rs.getString("username") + "\n");
+                ps.close();
+                connection.close();
+                return true;
+            }
+            else {
+                System.out.println("The user or the password doesn't match!");
+                ps.close();
+                connection.close();
+                return false;
+            }
+        }
+        catch (SQLException e)
+        {
+            System.out.println(e.getMessage());
+            return false;
+        }
     }
 }
