+++ untitled/src/viewer/LogInViewer.form
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="viewer.LogInViewer">
-  <grid id="27dc6" binding="LoginFrame" layout-manager="GridLayoutManager" row-count="4" column-count="39" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
+  <grid id="27dc6" binding="LoginFrame" layout-manager="GridLayoutManager" row-count="5" column-count="41" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
     <margin top="0" left="0" bottom="0" right="0"/>
     <constraints>
-      <xy x="20" y="20" width="1228" height="509"/>
+      <xy x="20" y="20" width="1266" height="509"/>
     </constraints>
     <properties>
       <background color="-262145"/>
@@ -12,7 +12,7 @@
     <children>
       <component id="93f33" class="javax.swing.JLabel">
         <constraints>
-          <grid row="0" column="0" row-span="1" col-span="39" vsize-policy="0" hsize-policy="0" anchor="0" fill="0" indent="0" use-parent-layout="false"/>
+          <grid row="0" column="0" row-span="1" col-span="41" vsize-policy="0" hsize-policy="0" anchor="0" fill="0" indent="0" use-parent-layout="false"/>
         </constraints>
         <properties>
           <font name="Arial Black" size="36" style="1"/>
@@ -22,28 +22,9 @@
           <verticalTextPosition value="3"/>
         </properties>
       </component>
-      <component id="6b765" class="javax.swing.JTextField" binding="Username">
-        <constraints>
-          <grid row="1" column="15" row-span="1" col-span="2" vsize-policy="0" hsize-policy="6" anchor="6" fill="0" indent="0" use-parent-layout="false">
-            <preferred-size width="150" height="-1"/>
-          </grid>
-        </constraints>
-        <properties>
-          <text value=""/>
-        </properties>
-      </component>
-      <component id="36aaf" class="javax.swing.JLabel" binding="Label">
-        <constraints>
-          <grid row="1" column="17" row-span="1" col-span="4" vsize-policy="0" hsize-policy="0" anchor="10" fill="0" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <font size="14" style="1"/>
-          <text value="שם משתמש"/>
-        </properties>
-      </component>
       <component id="ae2c5" class="javax.swing.JPasswordField" binding="password">
         <constraints>
-          <grid row="2" column="15" row-span="1" col-span="2" vsize-policy="0" hsize-policy="6" anchor="4" fill="0" indent="0" use-parent-layout="false">
+          <grid row="2" column="17" row-span="2" col-span="2" vsize-policy="0" hsize-policy="6" anchor="4" fill="0" indent="0" use-parent-layout="false">
             <preferred-size width="150" height="-1"/>
           </grid>
         </constraints>
@@ -53,7 +34,7 @@
       </component>
       <component id="534f7" class="javax.swing.JLabel" binding="pass">
         <constraints>
-          <grid row="2" column="17" row-span="1" col-span="1" vsize-policy="0" hsize-policy="0" anchor="0" fill="0" indent="0" use-parent-layout="false"/>
+          <grid row="2" column="19" row-span="2" col-span="1" vsize-policy="0" hsize-policy="0" anchor="0" fill="0" indent="0" use-parent-layout="false"/>
         </constraints>
         <properties>
           <font size="14" style="1"/>
@@ -62,7 +43,7 @@
       </component>
       <component id="8e222" class="javax.swing.JButton" binding="LogIn">
         <constraints>
-          <grid row="3" column="16" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="0" fill="0" indent="23" use-parent-layout="false"/>
+          <grid row="4" column="18" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="0" fill="0" indent="23" use-parent-layout="false"/>
         </constraints>
         <properties>
           <text value="התחבר"/>
@@ -70,7 +51,7 @@
       </component>
       <component id="9312" class="javax.swing.JButton" binding="register">
         <constraints>
-          <grid row="3" column="15" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="4" fill="0" indent="0" use-parent-layout="false"/>
+          <grid row="4" column="17" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="4" fill="0" indent="0" use-parent-layout="false"/>
         </constraints>
         <properties>
           <text value="עדיין לא נרשמת?"/>
@@ -78,14 +59,42 @@
       </component>
       <hspacer id="15ab3">
         <constraints>
-          <grid row="2" column="37" row-span="1" col-span="1" vsize-policy="1" hsize-policy="6" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
+          <grid row="2" column="39" row-span="2" col-span="1" vsize-policy="1" hsize-policy="6" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
         </constraints>
       </hspacer>
       <hspacer id="633a6">
         <constraints>
-          <grid row="3" column="0" row-span="1" col-span="1" vsize-policy="1" hsize-policy="6" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
+          <grid row="4" column="0" row-span="1" col-span="1" vsize-policy="1" hsize-policy="6" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
         </constraints>
       </hspacer>
+      <component id="f78c1" class="javax.swing.JLabel" binding="connect">
+        <constraints>
+          <grid row="3" column="18" row-span="1" col-span="2" vsize-policy="0" hsize-policy="0" anchor="0" fill="0" indent="0" use-parent-layout="false"/>
+        </constraints>
+        <properties>
+          <font name="Comic Sans MS" size="24" style="3"/>
+          <text value=""/>
+        </properties>
+      </component>
+      <component id="36aaf" class="javax.swing.JLabel" binding="Label">
+        <constraints>
+          <grid row="1" column="19" row-span="1" col-span="4" vsize-policy="0" hsize-policy="0" anchor="10" fill="0" indent="0" use-parent-layout="false"/>
+        </constraints>
+        <properties>
+          <font size="14" style="1"/>
+          <text value="שם משתמש"/>
+        </properties>
+      </component>
+      <component id="6b765" class="javax.swing.JTextField" binding="Username">
+        <constraints>
+          <grid row="1" column="18" row-span="1" col-span="1" vsize-policy="0" hsize-policy="6" anchor="6" fill="0" indent="0" use-parent-layout="false">
+            <preferred-size width="150" height="-1"/>
+          </grid>
+        </constraints>
+        <properties>
+          <text value=""/>
+        </properties>
+      </component>
     </children>
   </grid>
 </form>
